Query a Spectra object using MassQL

Query a Spectra object using MassQL

# S4 method for Spectra
query(x, query = character(), ...)

Arguments

x

The Spectra object to query.

query

character(1) with the MassQL query.

...

currently ignored.

Value

Depending on the <type of data> part of the MassQL query.

Type of data

The <type of data> allows to define which data should be extracted from the selected spectra. MassQL defines type of data being MS1DATA or MS2DATA to retrieve MS1 or MS2 scans. In addition, functions can be applied to these to modify (e.g. sum) the data. SpectraQL supports:

*: select all data and return the data subset as a Spectra() object.

MS1DATA: return a Spectra() with all MS1 scans from the selected spectra.

MS2DATA: return a Spectra() with all MS2 scans from the selected spectra.

scaninfo(MS1DATA), scaninfo(MS2DATA): return the spectraData() of all selected spectra.

scansum(MS1DATA), scansum(MS2DATA): sum of the peak intensities of the selected spectra (TIC).

scannum(MS1DATA), scannum(MS2DATA): the scan number(s) of the selected spectra.

Conditions

Conditions define to which spectra the data set should be subsetted. Several conditions can be combined with "and" (case insensitive). The syntax for a condition is <condition> = <value>, e.g. MS2PROD = 144.1. Such conditions can be further refined by additional expressions that allow for example to define acceptable tolerances for m/z differences. SpectraQL supports (case insensitive):

RTMIN: minimum retention time (in seconds).

RTMAX: maximum retention time (in seconds).

SCANMIN: the minimum scan number (acquisition number).

SCANMAX: the maximum scan number (acquisition number).

CHARGE: the charge for MS2 spectra.

POLARITY: the polarity of the spectra (can be "positive", "negative", "pos" or "neg", case insensitive).

MS2PROD: allows to select MS2 spectra that contain a peak with particular m/z value(s). See below for examples.

MS2PREC: allows to select MS2 spectra with the defined precursor m/z value(s). See below for examples.

MS1MZ: allows to select MS1 spectra containing peak(s) with the defined m/z value(s).

All conditions involving m/z values allow to specify a mass accuracy using the optional fields TOLERANCEMZ and TOLERANCEPPM that define the absolute and m/z-relative acceptable difference in m/z values. One or both fields can be attached to a condition such as MS2PREC=100:TOLERANCEMZ=0.1:TOLERANCEPPM=20 to select for example all MS2 spectra with a precursor m/z equal to 100 accepting a difference of 0.1 and 20 ppm. Note that in contrast to MassQL, the default tolarance and ppm is 0 for all calls.

Filters

Filters subset the data within spectra, i.e. select which peaks within spectra should be retrieved. SpectraQL does not support filters yet.

Author

Andrea Vicini, Johannes Rainer

Examples

## Read a data file with MS1 and MS2 spectra
library(msdata)
library(Spectra)
#> Loading required package: S4Vectors
#> Loading required package: stats4
#> Loading required package: BiocGenerics
#> 
#> Attaching package: ‘BiocGenerics’
#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs
#> The following objects are masked from ‘package:base’:
#> 
#>     Filter, Find, Map, Position, Reduce, anyDuplicated, append,
#>     as.data.frame, basename, cbind, colnames, dirname, do.call,
#>     duplicated, eval, evalq, get, grep, grepl, intersect, is.unsorted,
#>     lapply, mapply, match, mget, order, paste, pmax, pmax.int, pmin,
#>     pmin.int, rank, rbind, rownames, sapply, setdiff, sort, table,
#>     tapply, union, unique, unsplit, which.max, which.min
#> 
#> Attaching package: ‘S4Vectors’
#> The following objects are masked from ‘package:base’:
#> 
#>     I, expand.grid, unname
#> Loading required package: BiocParallel
#> 
#> Attaching package: ‘Spectra’
#> The following object is masked from ‘package:ProtGenerics’:
#> 
#>     addProcessing
fls <- dir(system.file("TripleTOF-SWATH", package = "msdata"),
    full.names = TRUE)
sps_dda <- Spectra(fls[1L])

## Subset to spectra measured between 300 and 400 seconds
query(sps_dda, "QUERY * WHERE RTMIN = 300 AND RTMAX = 400")
#> MSn data (Spectra) with 971 spectra in a MsBackendMzR backend:
#>       msLevel     rtime scanIndex
#>     <integer> <numeric> <integer>
#> 1           1   300.123      2271
#> 2           2   300.395      2272
#> 3           1   300.533      2273
#> 4           1   300.654      2274
#> 5           1   300.775      2275
#> ...       ...       ...       ...
#> 967         1   399.277      3237
#> 968         1   399.406      3238
#> 969         1   399.527      3239
#> 970         2   399.798      3240
#> 971         1   399.936      3241
#>  ... 33 more variables/columns.
#> 
#> file(s):
#> PestMix1_DDA.mzML
#> Processing:
#>  Filter: select retention time [300..400] on MS level(s) 1 2 [Fri Feb 11 07:16:26 2022] 

## To select only MS1 or MS2 spectra use "MS1DATA" or "MS2DATA" instead
## or *. Note also that queries are case-insensitive.
query(sps_dda, "query ms1data where rtmin = 300 and rtmax = 400")
#> MSn data (Spectra) with 326 spectra in a MsBackendMzR backend:
#>       msLevel     rtime scanIndex
#>     <integer> <numeric> <integer>
#> 1           1   300.123      2271
#> 2           1   300.533      2273
#> 3           1   300.654      2274
#> 4           1   300.775      2275
#> 5           1   301.183      2277
#> ...       ...       ...       ...
#> 322         1   399.152      3236
#> 323         1   399.277      3237
#> 324         1   399.406      3238
#> 325         1   399.527      3239
#> 326         1   399.936      3241
#>  ... 33 more variables/columns.
#> 
#> file(s):
#> PestMix1_DDA.mzML
#> Processing:
#>  Filter: select retention time [300..400] on MS level(s) 1 2 [Fri Feb 11 07:16:26 2022]
#>  Filter: select MS level(s) 1 [Fri Feb 11 07:16:26 2022] 

## To select (MS2) spectra with a certain precursor m/z the MS2PREC condition
## can be used. Below we extract all spectra with a precursor m/z of 99.9
## accepting also a difference of 10ppm
query(sps_dda, "QUERY * WHERE MS2PREC = 99.967:TOLERANCEPPM=10")
#> MSn data (Spectra) with 1 spectra in a MsBackendMzR backend:
#>     msLevel     rtime scanIndex
#>   <integer> <numeric> <integer>
#> 1         2   884.843      7495
#>  ... 33 more variables/columns.
#> 
#> file(s):
#> PestMix1_DDA.mzML
#> Processing:
#>  Filter: select spectra with precursor m/z matching 99.967 [Fri Feb 11 07:16:26 2022] 

## It is also possible to specify multiple precursor m/z values:
query(sps_dda, "QUERY * WHERE MS2PREC = (99.967 OR 428.88):TOLERANCEPPM=10")
#> MSn data (Spectra) with 2 spectra in a MsBackendMzR backend:
#>     msLevel     rtime scanIndex
#>   <integer> <numeric> <integer>
#> 1         2   884.843      7495
#> 2         2   887.813      7516
#>  ... 33 more variables/columns.
#> 
#> file(s):
#> PestMix1_DDA.mzML
#> Processing:
#>  Filter: select spectra with precursor m/z matching 99.967, 428.88 [Fri Feb 11 07:16:26 2022] 

## To select all MS1 spectra that contain a peak with a certain m/z we can
## use the MS1MZ condition. Below we combine this with an absolute tolerance
## using TOLERANCEMZ.
query(sps_dda, "QUERY * WHERE MS1MZ = 100:TOLERANCEMZ=1")
#> MSn data (Spectra) with 4625 spectra in a MsBackendMzR backend:
#>        msLevel     rtime scanIndex
#>      <integer> <numeric> <integer>
#> 1            1     0.231         1
#> 2            1     0.351         2
#> 3            1     0.471         3
#> 4            1     0.591         4
#> 5            1     0.711         5
#> ...        ...       ...       ...
#> 4621         1   899.491      7598
#> 4622         1   899.613      7599
#> 4623         1   899.747      7600
#> 4624         1   899.872      7601
#> 4625         1   899.993      7602
#>  ... 33 more variables/columns.
#> 
#> file(s):
#> PestMix1_DDA.mzML
#> Processing:
#>  Filter: select MS level(s) 1 [Fri Feb 11 07:16:26 2022] 

## Using MS2DATA in combination with MS1MZ will not return any spectra.
query(sps_dda, "QUERY MS2DATA WHERE MS1MZ = 100:TOLERANCEMZ=1")
#> MSn data (Spectra) with 0 spectra in a MsBackendMzR backend:
#> Processing:
#>  Filter: select MS level(s) 1 [Fri Feb 11 07:16:27 2022]
#>  Filter: select MS level(s) 2 [Fri Feb 11 07:16:27 2022] 

## In contrast, do select MS2 spectra containing a peak with a certain m/z
## we have to use the condition MS2PROD
query(sps_dda, "QUERY * WHERE MS2PROD = 100:TOLERANCEMZ=1")
#> MSn data (Spectra) with 260 spectra in a MsBackendMzR backend:
#>       msLevel     rtime scanIndex
#>     <integer> <numeric> <integer>
#> 1           2    47.183       370
#> 2           2    79.371       617
#> 3           2    81.661       631
#> 4           2    97.140       744
#> 5           2    97.550       746
#> ...       ...       ...       ...
#> 256         2   884.163      7491
#> 257         2   884.843      7495
#> 258         2   890.443      7535
#> 259         2   893.072      7554
#> 260         2   898.602      7591
#>  ... 33 more variables/columns.
#> 
#> file(s):
#> PestMix1_DDA.mzML
#> Processing:
#>  Filter: select MS level(s) 2 [Fri Feb 11 07:16:27 2022]